+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 This file contains highlights of what changes on each version of the [Coveralls for PHP](https://github.com/cedx/coveralls.php) package.
 
+## Version 6.0.0
+
 ## Version 5.0.0
 - Breaking change: moved the `Observable` API to a synchronous one.
 - Breaking change: moved the `Subject` event API to the `EventEmitter` one.